Move env.py out of commands dir.
