--- test/runner/lib/commands/__init__.py
@@ -1,1 +1,0 @@
-"""Commands for ansible-test."""
